test(e2e): re-enable mascot-context, scope desktop-selection (toward #114) (#320)
* test(e2e): fix mascot-tap and desktop-context-menu specs for prod build

- mascot: use a real Playwright click, not a synthetic PointerEvent — the
  handler calls setPointerCapture(pointerId) which throws for dispatched
  events, aborting the tap. Poll-while-closed since the tap toggles the chat.
- desktop-selection: scope the 'Terminal' locator to the new
  desktop-context-menu testid (an unscoped match hit two elements).

* test(e2e): scope desktop-selection Terminal locator, keep it fixme'd

The strict-mode 'two Terminal buttons' issue is fixed (scoped to the new
desktop-context-menu testid + retry), but launching the terminal still fails
against the bare standalone e2e server, which has no /terminal-ws proxy. Left
fixme'd with that diagnosis. Tracked in #114.

* test(e2e): extract terminal WS mock helper; diagnose desktop-selection

- Add e2e/helpers/mock-backends.ts (mockTerminalWebSocket) and use it from
  terminal-reconnect (DRY; validated still green on the Jetson prod build).
- desktop-selection: locator scoping done, WS mock wired, but re-fixme'd with a
  full diagnosis — the (40,40) right-click doesn't open the desktop context
  menu on the prod build (hit-testing/layout), which must be solved first.
  Tracked in #114.

* test(e2e): re-enable browser-vnc, clawkeep unpair, installed-app-settings

Fix three more #114 specs against the production build:
- browser-vnc: the launcher paginates apps; search for 'Browser' instead of
  relying on page order.
- clawkeep-interactions (unpair): the Unpair button sits under the taskbar;
  dispatch the click to the element so hit-testing doesn't land on the shelf.
- installed-app-settings: the freshly-installed icon animates in and never
  passes the stability gate; dispatch the click to its launch button.

desktop-selection: re-fixme'd with a precise finding — its context-menu handler
fires but the menu never renders on the standalone production build (every input
path), which looks like a real app-side bug, not a test issue. Adds a
desktop-surface testid as groundwork. Tracked in #114.
